Remove BlockInput case postEIP4844OldBlobs
